♻️ add gh commands
